+++ scripts/data/equipment_data.gd
@@ -2,7 +2,8 @@ class_name EquipmentData
 extends Resource
 ## Pezzo di equipaggiamento permanente. Il bilanciamento si fa solo nei .tres in res://data/equipment/.
 
-enum Slot { WEAPON, ACCESSORY }
+## Tipo di oggetto (M11). ACCESSORY = Amuleto. Nuove voci solo in coda (valore salvato nei .tres).
+enum Slot { WEAPON, ACCESSORY, HEAD, GLOVES, ARMOR, PANTS, BOOTS, RING }
 
 @export var id: StringName = &""
 @export var display_name: String = ""